+++ polaris.shopify.com/content/components/overlays/hovercard.mdx
@@ -1,0 +1,39 @@
+---
+title: HoverCard
+shortDescription: Used to present a preview of a commerce object's key information when hovering a link to its detail page.
+category: Overlays
+keywords:
+  - hovercard
+  - link preview
+  - link details
+  - commerce object
+examples:
+  - fileName: hovercard-with-child-activator.tsx
+    title: With child activator
+    description: A hover card that renders and is triggered to be `active` by its `children`. Use for commerce objects rendered by themselves within normal page content.
+  - fileName: hovercard-with-dynamic-activator.tsx
+    title: With dynamic activator
+    description: A HoverCard rendered without `children` that is triggered to be `active` and repositioned by dynamically setting the `activator` prop. Use when several of the same commerce object type are in close context, like a column of customer names in an index table of orders.
+---
+
+# {frontmatter.title}
+
+<Lede>
+
+A hover card is an overlay only triggered by mouse over of a link. They are not triggered on focus, keyboard navigable, or visible to screen readers. Use to present a preview of a commerce object's key information when hovering a link to its detail page.
+
+</Lede>
+
+<Examples />
+
+<Props componentName={frontmatter.title} />
+
+## Best practices
+
+---
+
+## Content guidelines
+
+---
+
+## Related components
