+++ polaris.shopify.com/pages/examples/hovercard-with-dynamic-activator.tsx
@@ -1,13 +1,9 @@
-import React, {useCallback, useState, useRef} from 'react';
-import type {ComponentMeta} from '@storybook/react';
-import type {PositionedOverlayProps} from '@shopify/polaris';
+import React, {useCallback, useState} from 'react';
 import {
   Tag,
   Thumbnail,
   useIndexResourceState,
-  ButtonGroup,
   Box,
-  Button,
   HoverCard,
   useHoverCardActivatorWrapperProps,
   Badge,
@@ -25,129 +21,9 @@ import {
   OrdersMinor,
   ImageMajor,
 } from '@shopify/polaris-icons';
+import {withPolarisExample} from '../../src/components/PolarisExampleWrapper';
 
-export default {
-  component: HoverCard,
-} as ComponentMeta<typeof HoverCard>;
-
-export function WithChildActivator() {
-  const [active, setActive] = useState(false);
-  const [position, setPosition] =
-    useState<PositionedOverlayProps['preferredPosition']>('below');
-
-  const handleChangePosition =
-    (position: PositionedOverlayProps['preferredPosition']) => () => {
-      setPosition(position);
-    };
-
-  const activator = (
-    <Link removeUnderline url="#">
-      Saul Goodman
-    </Link>
-  );
-
-  const customerHoverCardContent = (
-    <Box padding="400">
-      <BlockStack gap="400">
-        <BlockStack gap="0">
-          <Text as="span" variant="headingSm">
-            <Link removeUnderline>Saul Goodman</Link>
-          </Text>
-          <Text as="span" variant="bodyMd">
-            <Link url="mailto:help@bettercallsaul.com">
-              help@bettercallsaul.com
-            </Link>
-          </Text>
-          <Text as="p" variant="bodyMd">
-            <Link url="tel:+1505-842-5662">+1 505-842-5662</Link>
-          </Text>
-        </BlockStack>
-        <Box width="100%">
-          <BlockStack gap="100">
-            <InlineStack wrap={false} gap="100" align="start">
-              <Icon tone="subdued" source={LocationsMinor} />
-              <Text tone="subdued" as="p">
-                Albequerque, NM, USA
-              </Text>
-            </InlineStack>
-            <InlineStack wrap={false} gap="100" align="start">
-              <Box>
-                <Icon tone="subdued" source={OrdersMinor} />
-              </Box>
-              <Text tone="subdued" as="p">
-                8 Orders
-              </Text>
-            </InlineStack>
-          </BlockStack>
-        </Box>
-      </BlockStack>
-    </Box>
-  );
-
-  const positionControlBar = (
-    <BlockStack gap="100" inlineAlign="center">
-      <Box id="Filler" minHeight="200px" />
-      <Text as="p" tone="subdued">
-        Use the buttons below to change the hover card position
-      </Text>
-      <ButtonGroup variant="segmented">
-        <Button
-          pressed={position === 'left'}
-          onClick={handleChangePosition('left')}
-        >
-          Left
-        </Button>
-        <Button
-          pressed={position === 'right'}
-          onClick={handleChangePosition('right')}
-        >
-          Right
-        </Button>
-        <Button
-          pressed={position === 'above'}
-          onClick={handleChangePosition('above')}
-        >
-          Above
-        </Button>
-        <Button
-          pressed={position === 'below'}
-          onClick={handleChangePosition('below')}
-        >
-          Below
-        </Button>
-      </ButtonGroup>
-    </BlockStack>
-  );
-
-  return (
-    <div style={{margin: '0 auto', height: '600px', width: '960px'}}>
-      <InlineStack align="center" blockAlign="center">
-        <BlockStack gap="500">
-          {positionControlBar}
-
-          <Card>
-            <BlockStack gap="300">
-              <Text as="h2" variant="headingSm">
-                Customer
-              </Text>
-              <HoverCard
-                active={active}
-                preferredPosition={position}
-                preferredAlignment="left"
-                content={customerHoverCardContent}
-                toggleActive={setActive}
-              >
-                {activator}
-              </HoverCard>
-            </BlockStack>
-          </Card>
-        </BlockStack>
-      </InlineStack>
-    </div>
-  );
-}
-
-export function WithDynamicActivator() {
+export function HoverCardWithDynamicActivator() {
   interface CustomerDetailPreview {
     id: string;
     name: string;
@@ -337,7 +213,7 @@ export function WithDynamicActivator() {
     activatorElement,
     handleMouseEnterActivator,
     handleMouseLeaveActivator,
-  } = useHoverCardActivatorWrapperProps({snapToParent: true, hoverDelay: 100});
+  } = useHoverCardActivatorWrapperProps({snapToParent: true});
 
   const handleMouseEnterCustomer = useCallback(
     (customer: CustomerDetailPreview) =>
@@ -612,38 +488,43 @@ export function WithDynamicActivator() {
   );
 
   return (
-    <Card padding="0">
-      <HoverCard
-        snapToParent
-        minWidth={activeHoverCard.order ? 300 : undefined}
-        activator={activatorElement}
-        activatorWrapper="div"
-        preferredPosition="right"
-        content={hoverCardContent}
-      />
-      <IndexTable
-        resourceName={resourceName}
-        itemCount={orders.length}
-        selectedItemsCount={
-          allResourcesSelected ? 'All' : selectedResources.length
-        }
-        onSelectionChange={handleSelectionChange}
-        headings={[
-          {title: 'Order'},
-          {title: 'Date'},
-          {title: 'Customer'},
-          {title: 'Channel'},
-          {title: 'Total', alignment: 'end'},
-          {title: 'Payment status'},
-          {title: 'Fulfillment status'},
-          {title: 'Items'},
-          {title: 'Delivery status'},
-          {title: 'Delivery method'},
-          {title: 'Tags'},
-        ]}
-      >
-        {rowMarkup}
-      </IndexTable>
-    </Card>
+    <>
+      <Card padding="0">
+        <HoverCard
+          snapToParent
+          minWidth={activeHoverCard.order ? 300 : undefined}
+          activator={activatorElement}
+          activatorWrapper="div"
+          preferredPosition="right"
+          content={hoverCardContent}
+        />
+        <IndexTable
+          resourceName={resourceName}
+          itemCount={orders.length}
+          selectedItemsCount={
+            allResourcesSelected ? 'All' : selectedResources.length
+          }
+          onSelectionChange={handleSelectionChange}
+          headings={[
+            {title: 'Order'},
+            {title: 'Date'},
+            {title: 'Customer'},
+            {title: 'Channel'},
+            {title: 'Total', alignment: 'end'},
+            {title: 'Payment status'},
+            {title: 'Fulfillment status'},
+            {title: 'Items'},
+            {title: 'Delivery status'},
+            {title: 'Delivery method'},
+            {title: 'Tags'},
+          ]}
+        >
+          {rowMarkup}
+        </IndexTable>
+      </Card>
+      <Box minHeight="200px" />
+    </>
   );
 }
+
+export default withPolarisExample(HoverCardWithDynamicActivator);
